whisper-server/requirements.txt
openai-whisper==20231117
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
torch==2.0.1
torchaudio==2.0.2